Move network calls re accounts to account list
